Split stuff out of main.c into more files.
Fixed issue with player input.
