+++ Dark Samurai Warrior Reloaded/math.h
@@ -14,7 +14,7 @@ typedef union V4 {
   };
 } V4;
 
-V4 v4(float r, float g, float b, float a) {
+inline V4 v4(float r, float g, float b, float a) {
   assert(r <= 1.0f && r >= 0.0f);
   assert(g <= 1.0f && g >= 0.0f);
   assert(b <= 1.0f && b >= 0.0f);
