+++ Dark Samurai Warrior Reloaded/render.h
@@ -1,0 +1,7 @@
+#pragma once
+typedef struct LoadedBitmap {
+  int width;
+  int height;
+  int pitch;
+  void *memory;
+} LoadedBitmap;
